+++ bitfunk-version-catalog-accessor/src/main/kotlin/eu/bitfunk/gradle/version/catalog/intern/ResourceLoader.kt
@@ -16,18 +16,21 @@
  * PERFORMANCE OF THIS SOFTWARE.
  */
 
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        mavenCentral()
-        google()
+package eu.bitfunk.gradle.version.catalog.intern
+
+import java.io.IOException
+import java.nio.charset.StandardCharsets
+
+internal object ResourceLoader : InternalContract.ResourceLoader {
+    private val FILE_ENCODING = StandardCharsets.UTF_8
+
+    @Suppress("RECEIVER_NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS")
+    @Throws(IOException::class)
+    override fun loadAsString(filePath: String): String {
+        try {
+            return this::class.java.classLoader.getResource(filePath).readText(FILE_ENCODING)
+        } catch (error: Throwable) {
+            throw NullPointerException("$filePath does not exist")
+        }
     }
-
-    includeBuild("bitfunk-version-catalog-accessor")
 }
-
-include("docs")
-
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-
-rootProject.name = "GradlePlugins"