+++ bitfunk-version-catalog-accessor/src/test/resources/fixture/EmptyVersionCatalogAccessor.kt
@@ -1,0 +1,18 @@
+package com.example.catalog
+
+import eu.bitfunk.gradle.version.catalog.accessor.BaseVersionCatalogAccessor
+import eu.bitfunk.gradle.version.catalog.accessor.VersionCatalogDependency
+import org.gradle.api.Project
+
+public class EmptyVersionCatalogAccessor(
+    project: Project
+) : BaseVersionCatalogAccessor(project, "empty") {
+    public val versions: VersionCatalogDependency.Group = object : VersionCatalogDependency.Group {
+    }
+
+    public val bundles: VersionCatalogDependency.Group = object : VersionCatalogDependency.Group {
+    }
+
+    public val plugins: VersionCatalogDependency.Group = object : VersionCatalogDependency.Group {
+    }
+}